[api-rewrite] working post /request
