+++ lib/sensu/api/routes.rb
@@ -2,6 +2,7 @@ require "sensu/api/routes/info"
 require "sensu/api/routes/health"
 require "sensu/api/routes/clients"
 require "sensu/api/routes/checks"
+require "sensu/api/routes/request"
 
 module Sensu
   module API
@@ -10,6 +11,7 @@ module Sensu
       include Health
       include Clients
       include Checks
+      include Request
     end
   end
 end
